create CameraX Binding file
