+++ app/src/main/java/com/example/vtubercamera/ui/screens/CameraScreen.kt
@@ -84,6 +84,7 @@ import com.example.vtubercamera.ui.components.PermissionRequestComponent
 import com.example.vtubercamera.ui.components.PhotoDetailView
 import com.example.vtubercamera.ui.components.PhotoPreviewComponent
 import com.example.vtubercamera.ui.modifiers.modernCameraGestures
+import com.example.vtubercamera.ui.camerax.bindCameraWithPreview
 import com.example.vtubercamera.ui.viewmodels.CameraViewModel
 import com.example.vtubercamera.utils.CameraCapabilityManager
 import com.example.vtubercamera.utils.PermissionUtils
@@ -939,57 +940,3 @@ private fun ThumbnailPreview(
     }
 }
 
-
-private fun bindCameraWithPreview(
-    lifecycleOwner: LifecycleOwner,
-    cameraProvider: ProcessCameraProvider,
-    preview: Preview,
-    cameraSelector: CameraSelector,
-    flashMode: Int,
-    initialZoomRatio: Float? = null,
-    onImageCaptureCreated: (ImageCapture) -> Unit,
-    onCameraCreated: (Camera) -> Unit
-): Camera? {
-    return try {
-        Log.d(
-            "CameraBinding",
-            "Binding camera with flash mode: $flashMode, camera: ${if (cameraSelector == CameraSelector.DEFAULT_BACK_CAMERA) "BACK" else "FRONT"}"
-        )
-
-        val imageCapture = ImageCapture.Builder()
-            .setCaptureMode(ImageCapture.CAPTURE_MODE_MAXIMIZE_QUALITY)
-            .setFlashMode(flashMode)
-            .build()
-
-        cameraProvider.unbindAll()
-
-        val camera = cameraProvider.bindToLifecycle(
-            lifecycleOwner,
-            cameraSelector,
-            preview,
-            imageCapture
-        )
-
-        onImageCaptureCreated(imageCapture)
-        onCameraCreated(camera)
-
-        initialZoomRatio?.let { targetZoom ->
-            try {
-                val minZoom = camera.cameraInfo.zoomState.value?.minZoomRatio ?: 1.0f
-                val maxZoom = camera.cameraInfo.zoomState.value?.maxZoomRatio ?: 10.0f
-                val coerced = targetZoom.coerceIn(minZoom, maxZoom)
-                camera.cameraControl.setZoomRatio(coerced)
-                Log.d("CameraBinding", "Restored zoom ratio to $coerced after rebind")
-            } catch (e: Exception) {
-                Log.w("CameraBinding", "ズーム復元に失敗しました", e)
-            }
-        }
-
-        Log.d("CameraBinding", "Camera bound successfully")
-
-        camera
-    } catch (e: Exception) {
-        Log.e("CameraBinding", "カメラのバインドに失敗しました", e)
-        null
-    }
-}
